# Optimise relabeling by re-using memory