--- a/model/labels/labels.go
+++ b/model/labels/labels.go
@@ -467,17 +467,25 @@ func (b *Builder) Set(n, v string) *Builder {
 	return b
 }
 
-// Labels returns the labels from the builder. If no modifications
-// were made, the original labels are returned.
-func (b *Builder) Labels() Labels {
+// Labels returns the labels from the builder, adding them to res if non-nil.
+// Argument res can be the same as b.base, if caller wants to overwrite that slice.
+// If no modifications were made, the original labels are returned.
+func (b *Builder) Labels(res Labels) Labels {
 	if len(b.del) == 0 && len(b.add) == 0 {
 		return b.base
 	}
 
-	// In the general case, labels are removed, modified or moved
-	// rather than added.
-	res := make(Labels, 0, len(b.base))
+	if res == nil {
+		// In the general case, labels are removed, modified or moved
+		// rather than added.
+		res = make(Labels, 0, len(b.base))
+	} else {
+		res = res[:0]
+	}
 Outer:
+	// Justification that res can be the same slice as base: in this loop
+	// we move forward through base, and either skip an element or assign
+	// it to res at its current position or an earlier position.
 	for _, l := range b.base {
 		for _, n := range b.del {
 			if l.Name == n {