--- a/model/labels/labels_test.go
+++ b/model/labels/labels_test.go
@@ -722,7 +722,7 @@ func TestBuilder(t *testing.T) {
 				b.Keep(tcase.keep...)
 			}
 			b.Del(tcase.del...)
-			require.Equal(t, tcase.want, b.Labels())
+			require.Equal(t, tcase.want, b.Labels(tcase.base))
 		})
 	}
 }