--- a/model/relabel/relabel.go
+++ b/model/relabel/relabel.go
@@ -194,15 +194,8 @@ func (re Regexp) String() string {
 // May return the input labelSet modified.
 func Process(lbls labels.Labels, cfgs ...*Config) labels.Labels {
 	lb := labels.NewBuilder(nil)
-	max := 0
 	for _, cfg := range cfgs {
-		if len(cfg.SourceLabels) > max {
-			max = len(cfg.SourceLabels)
-		}
-	}
-	values := make([]string, 0, max)
-	for _, cfg := range cfgs {
-		lbls = relabel(lbls, cfg, lb, values[:0])
+		lbls = relabel(lbls, cfg, lb)
 		if lbls == nil {
 			return nil
 		}
@@ -210,7 +203,12 @@ func Process(lbls labels.Labels, cfgs ...*Config) labels.Labels {
 	return lbls
 }
 
-func relabel(lset labels.Labels, cfg *Config, lb *labels.Builder, values []string) labels.Labels {
+func relabel(lset labels.Labels, cfg *Config, lb *labels.Builder) labels.Labels {
+	var va [16]string
+	var values []string = va[:0]
+	if len(cfg.SourceLabels) > cap(values) {
+		values = make([]string, 0, len(cfg.SourceLabels))
+	}
 	for _, ln := range cfg.SourceLabels {
 		values = append(values, lset.Get(string(ln)))
 	}