Relabeling can be a large portion of Prometheus memory allocation, and hence overall performance.
Especially relabeling in service discovery.

Since relabelling works by applying different rules to the same label set, we can re-use slices and greatly reduce memory allocation.  Also re-use the Builder itself since `NewBuilder()` makes some speculative allocations.

The benchmark is somewhat lengthy, due to including realistic data.

```
name                  old time/op    new time/op    delta
Relabel/example-4       5.32µs ± 4%    4.10µs ± 2%  -22.92%  (p=0.008 n=5+5)
Relabel/kubernetes-4    50.0µs ± 7%    33.6µs ± 2%  -32.93%  (p=0.008 n=5+5)

name                  old alloc/op   new alloc/op   delta
Relabel/example-4       1.83kB ± 0%    0.74kB ± 0%  -59.66%  (p=0.000 n=4+5)
Relabel/kubernetes-4    26.3kB ± 0%     3.4kB ± 0%  -87.07%  (p=0.008 n=5+5)

name                  old allocs/op  new allocs/op  delta
Relabel/example-4         33.0 ± 0%      23.0 ± 0%  -30.30%  (p=0.008 n=5+5)
Relabel/kubernetes-4       106 ± 0%        44 ± 0%  -58.49%  (p=0.008 n=5+5)
```