@bboreham 2022-08-12
I pushed another commit which re-does "[re-use source values slice](https://github.com/prometheus/prometheus/pull/11147/commits/1652a4586b58ec30ad806347f910533907ae5225)" in a simpler way, and reduces allocations by 1 more.